+++ packages/components/src/Form/Fields/FieldRadio/FieldRadio.story.tsx
@@ -26,81 +26,48 @@
 
 import React from 'react'
 import { Story } from '@storybook/react/types-6-0'
-import { IconButton, IconButtonProps } from './IconButton'
+import { FieldRadio, FieldRadioProps } from './FieldRadio'
 
-export default {
-  component: IconButton,
-  title: 'IconButton',
-}
-
-const Template: Story<IconButtonProps> = (args) => <IconButton {...args} />
+const Template: Story<FieldRadioProps> = (args) => <FieldRadio {...args} />
 
 export const Basic = Template.bind({})
 Basic.args = {
-  icon: 'Plus',
-  label: 'Add',
+  id: 'fieldRadioId',
+  label: 'Field Radio Example',
+  name: 'thumbsUp',
 }
 
-export const XXSmall = Template.bind({})
-XXSmall.args = {
+export const Detail = Template.bind({})
+Detail.args = {
   ...Basic.args,
-  size: 'xxsmall',
+  detail: 'Some exciting details that describe the use of this.',
 }
 
-export const XSmall = Template.bind({})
-XSmall.args = {
+export const Checked = Template.bind({})
+Checked.args = {
   ...Basic.args,
-  size: 'xsmall',
+  checked: true,
 }
 
-export const Small = Template.bind({})
-Small.args = {
+export const Disabled = Template.bind({})
+Disabled.args = {
   ...Basic.args,
-  size: 'small',
+  disabled: true,
 }
 
-export const Large = Template.bind({})
-Large.args = {
-  ...Basic.args,
-  size: 'large',
+export const DisabledChecked = Template.bind({})
+DisabledChecked.args = {
+  ...Checked.args,
+  disabled: true,
 }
 
-export const OutlineXXSmall = Template.bind({})
-OutlineXXSmall.args = {
+export const ReadOnly = Template.bind({})
+ReadOnly.args = {
   ...Basic.args,
-  outline: true,
-  size: 'xxsmall',
+  readOnly: true,
 }
 
-export const OutlineXSmall = Template.bind({})
-OutlineXSmall.args = {
-  ...Basic.args,
-  outline: true,
-  size: 'xsmall',
-}
-
-export const OutlineSmall = Template.bind({})
-OutlineSmall.args = {
-  ...Basic.args,
-  outline: true,
-  size: 'small',
-}
-
-export const OutlineLarge = Template.bind({})
-OutlineLarge.args = {
-  ...Basic.args,
-  outline: true,
-  size: 'large',
-}
-
-export const ToggleOn = Template.bind({})
-ToggleOn.args = {
-  ...Basic.args,
-  toggle: true,
-}
-
-export const ToggleOff = Template.bind({})
-ToggleOff.args = {
-  ...Basic.args,
-  toggle: false,
+export default {
+  component: FieldRadio,
+  title: 'FieldRadio',
 }