+++ packages/components/src/InputDateRange/InputDateRange.story.tsx
@@ -49,8 +49,8 @@ Basic.args = {
 export const Disabled = Template.bind({})
 Disabled.args = {
   defaultValue: {
-    from: new Date('Oct 11, 2020'),
-    to: new Date('Oct 12, 2020'),
+    from: new Date('Jun 7, 2000'),
+    to: new Date('Jun 19, 2000'),
   },
   disabled: true,
 }