+++ packages/components/src/Menu/MenuItem.tsx
@@ -84,7 +84,6 @@ const MenuItemInternal: FC<MenuItemProps> = (props) => {
     onBlur,
     onClick,
     onKeyUp,
-    rel,
     target,
     tooltip,
     tooltipPlacement,
@@ -165,6 +164,18 @@ const MenuItemInternal: FC<MenuItemProps> = (props) => {
     )
   }
   const Component = !disabled && itemRole === 'link' ? 'a' : 'button'
+
+  /**
+   * `target="_blank" can be used to reverse tab-nab
+   * https://owasp.org/www-community/attacks/Reverse_Tabnabbing
+   */
+  const noTabNab = 'noopener noreferrer'
+  const rel =
+    target === '_blank'
+      ? props.rel
+        ? `${props.rel} ${noTabNab}`
+        : noTabNab
+      : props.rel
 
   const menuItemContent = (
     <Component href={href} rel={rel} role="menuitem" target={target}>